Feature events

+++ HokutosaiApp/NewsViewController.swift
@@ -10,8 +10,8 @@ import UIKit
 
 class NewsViewController: UIViewController, TappableViewControllerDelegate, UITableViewDelegate, UITableViewDataSource, LikeableTableViewCellDelegate, TabBarIntaractiveController {
     
-    private var topics: [TopicNews]!
-    private var articles: [Article]!
+    private var topics: [TopicNews]?
+    private var articles: [Article]?
     
     private var topicsBordController: FlowingPageViewController!
     private var timeline: UITableView!
@@ -52,8 +52,8 @@ class NewsViewController: UIViewController, TappableViewControllerDelegate, UITa
         // Dispose of any resources that can be recreated.
         
         if self.articles != nil {
-            while (UInt(self.articles.count) > self.onceGetArticleCount) {
-                self.articles.removeLast()
+            while (UInt(self.articles!.count) > self.onceGetArticleCount) {
+                self.articles!.removeLast()
             }
             self.articlesHitBottom = false
             self.timeline?.reloadData()
@@ -83,7 +83,7 @@ class NewsViewController: UIViewController, TappableViewControllerDelegate, UITa
         self.timeline.layoutMargins = UIEdgeInsetsZero
         self.timeline.separatorInset = UIEdgeInsetsZero
         
-        self.timeline.setContentAndScrollInsets(UIEdgeInsets(top: 0.0, left: 0.0, bottom: self.tabBarHeight, right: 0.0))
+        self.timeline.setContentAndScrollInsets(UIEdgeInsets(top: 0.0, left: 0.0, bottom: MainTabViewController.mainController.tabBar.height, right: 0.0))
         
         self.timeline.dataSource = self
         self.timeline.delegate = self
@@ -98,20 +98,19 @@ class NewsViewController: UIViewController, TappableViewControllerDelegate, UITa
         self.updatingTopics = true
         
         HokutosaiApi.GET(HokutosaiApi.News.Topics()) { response in
-            guard response.isSuccess else {
-                self.presentViewController(ErrorAlert.Server.failureGet(), animated: true, completion: nil)
+            guard response.isSuccess, let data = response.model else {
                 self.updatingTopics = false
                 completion?()
                 return
             }
             
-            self.topics = response.model
+            self.topics = data
             
             var pages = [TopicViewController]()
-            for i in 0 ..< self.topics.count {
+            for i in 0 ..< data.count {
                 let topicViewController = TopicViewController()
                 topicViewController.view.frame = CGRect(x: 0.0, y: 0.0, width: self.topicsBordController.viewSize.width, height: self.topicsBordController.viewSize.height)
-                topicViewController.setTopicContentData(i, data: self.topics[i])
+                topicViewController.setTopicContentData(i, data: data[i])
                 topicViewController.delegate = self
                 pages.append(topicViewController)
             }
@@ -141,9 +140,6 @@ class NewsViewController: UIViewController, TappableViewControllerDelegate, UITa
             guard response.isSuccess, let data = response.model else {
                 if lastId != nil {
                     self.loadingCellManager.status = .ReadyReload
-                }
-                else {
-                    self.presentViewController(ErrorAlert.Server.failureGet(), animated: true, completion: nil)
                 }
                 self.updatingTimeline = false
                 completion?()
@@ -198,10 +194,11 @@ class NewsViewController: UIViewController, TappableViewControllerDelegate, UITa
     }
     
     func tableView(tableView: UITableView, didSelectRowAtIndexPath indexPath: NSIndexPath) {
-        guard indexPath.row < self.articles.count else {
+        guard let articles = self.articles else { return }
+        guard indexPath.row < articles.count else {
             tableView.deselectRowAtIndexPath(indexPath, animated: true)
             self.loadingCellManager.status = .Loading
-            self.updateTimeline(self.articles.last?.newsId)
+            self.updateTimeline(articles.last?.newsId)
             return
         }
         
@@ -222,7 +219,8 @@ class NewsViewController: UIViewController, TappableViewControllerDelegate, UITa
     }
     
     func tableView(tableView: UITableView, heightForRowAtIndexPath indexPath: NSIndexPath) -> CGFloat {
-        guard indexPath.row < self.articles.count else {
+        guard let articles = self.articles else { return NewsTableViewCell.rowHeight }
+        guard indexPath.row < articles.count else {
             return LoadingCellManager.cellRowHeight
         }
         
@@ -234,21 +232,23 @@ class NewsViewController: UIViewController, TappableViewControllerDelegate, UITa
     }
     
     func tableView(tableView: UITableView, cellForRowAtIndexPath indexPath: NSIndexPath) -> UITableViewCell {
-        guard indexPath.row < self.articles.count else {
-            self.updateTimeline(self.articles.last?.newsId)
+        guard indexPath.row < self.articles!.count else {
+            self.updateTimeline(self.articles!.last?.newsId)
             return self.loadingCellManager.cell
         }
         
         let cell = tableView.dequeueReusableCellWithIdentifier(cellIdentifier, forIndexPath: indexPath) as! NewsTableViewCell
         
-        cell.changeData(indexPath.row, article: self.articles[indexPath.row])
+        cell.changeData(indexPath.row, article: self.articles![indexPath.row])
         cell.delegate = self
         
         return cell
     }
     
     func like(index: Int, cell: LikeableTableViewCell) {
-        let newsId = self.articles[index].newsId!
+        guard self.articles != nil else { return }
+        
+        let newsId = self.articles![index].newsId!
         HokutosaiApi.POST(HokutosaiApi.News.Likes(newsId: newsId)) { response in
             guard let result = response.model else {
                 self.presentViewController(ErrorAlert.Server.failureSendRequest(), animated: true, completion: nil)
@@ -256,14 +256,16 @@ class NewsViewController: UIViewController, TappableViewControllerDelegate, UITa
                 return
             }
             
-            self.articles[index].liked = result.liked
-            self.articles[index].likesCount = result.likesCount
+            self.articles![index].liked = result.liked
+            self.articles![index].likesCount = result.likesCount
             cell.updateLikes(newsId)
         }
     }
     
     func dislike(index: Int, cell: LikeableTableViewCell) {
-        let newsId = self.articles[index].newsId!
+        guard self.articles != nil else { return }
+        
+        let newsId = self.articles![index].newsId!
         HokutosaiApi.DELETE(HokutosaiApi.News.Likes(newsId: newsId)) { response in
             guard let result = response.model else {
                 self.presentViewController(ErrorAlert.Server.failureSendRequest(), animated: true, completion: nil)
@@ -271,8 +273,8 @@ class NewsViewController: UIViewController, TappableViewControllerDelegate, UITa
                 return
             }
             
-            self.articles[index].liked = result.liked
-            self.articles[index].likesCount = result.likesCount
+            self.articles![index].liked = result.liked
+            self.articles![index].likesCount = result.likesCount
             cell.updateLikes(newsId)
         }
     }
@@ -290,6 +292,7 @@ class NewsViewController: UIViewController, TappableViewControllerDelegate, UITa
     
     private let durationForOpenClose: NSTimeInterval = 0.3
     private let advancedDistanceThresholdForClosing: CGFloat = 50.0
+    private let contentLimitForClosing = 10
     private let speedThresholdForOpening: CGFloat = -30.0
     private let offsetThresholdForOpening: CGFloat = 100.0
     
@@ -302,11 +305,12 @@ class NewsViewController: UIViewController, TappableViewControllerDelegate, UITa
     }
     
     private func openCloseTopicsBord(offsetY: CGFloat) {
+        guard let articles = self.articles else { return }
         let advancedDistance = offsetY - self.scrollStartOffsetY
         
-        if self.topicsBordOpened && advancedDistance >= self.advancedDistanceThresholdForClosing {
+        if self.topicsBordOpened && advancedDistance >= self.advancedDistanceThresholdForClosing && articles.count >= self.contentLimitForClosing {
             self.topicsBordOpened = false
-            self.timeline.setContentAndScrollInsets(UIEdgeInsets(top: self.appearOriginY, left: 0.0, bottom: self.tabBarHeight, right: 0.0))
+            self.timeline.setContentAndScrollInsets(UIEdgeInsets(top: self.appearOriginY, left: 0.0, bottom: MainTabViewController.mainController.tabBar.height, right: 0.0))
             UIView.animateWithDuration(self.durationForOpenClose) {
                 self.topicsBordController.viewOrigin.y = -self.topicsBordController.viewSize.height
                 self.timeline.frame = self.timeLineFrame
@@ -319,7 +323,7 @@ class NewsViewController: UIViewController, TappableViewControllerDelegate, UITa
                 UIView.animateWithDuration(self.durationForOpenClose, animations: {
                     self.topicsBordController.viewOrigin.y = self.appearOriginY
                     }, completion: { _ in
-                        self.timeline.setContentAndScrollInsets(UIEdgeInsets(top: 0.0, left: 0.0, bottom: self.tabBarHeight, right: 0.0))
+                        self.timeline.setContentAndScrollInsets(UIEdgeInsets(top: 0.0, left: 0.0, bottom: MainTabViewController.mainController.tabBar.height, right: 0.0))
                         UIView.animateWithDuration(self.durationForOpenClose) {
                             self.timeline.frame = self.timeLineFrame
                         }
